+++ Realm/Tests/Swift/SwiftDynamicTests.swift
@@ -62,7 +62,7 @@ class SwiftDynamicTests: RLMTestCase {
 
         // verify properties
         let dyrealm = realm(withTestPathAndSchema: nil)
-        let array = dyrealm.allObjects("SwiftDynamicObject")
+        let array = dyrealm.allObjects("SwiftDynamicObject") as! RLMResults<RLMObject>
 
         XCTAssertTrue(array[0]["intCol"] as! NSNumber == 1)
         XCTAssertTrue(array[1]["stringCol"] as! String == "column2")
@@ -107,7 +107,7 @@ class SwiftDynamicTests: RLMTestCase {
 
         // verify properties
         let dyrealm = realm(withTestPathAndSchema: nil)
-        let array = dyrealm.allObjects("DynamicObject")
+        let array = dyrealm.allObjects("DynamicObject") as! RLMResults<RLMObject>
 
         XCTAssertTrue(array[0]["intCol"] as! NSNumber == 1)
         XCTAssertTrue(array[1]["stringCol"] as! String == "column2")
@@ -137,7 +137,7 @@ class SwiftDynamicTests: RLMTestCase {
 
         // verify properties
         let dyrealm = realm(withTestPathAndSchema: nil)
-        let results = dyrealm.allObjects(AllTypesObject.className())
+        let results = dyrealm.allObjects(AllTypesObject.className()) as! RLMResults<RLMObject>
         XCTAssertEqual(results.count, UInt(2))
         let robj1 = results[0]
         let robj2 = results[1]